Apply code review fixes
- Add retry queue to PostgresSettingsRepository.set() to prevent
  silent data loss on transient DB failures (#1)
- Make doctor.ts configDir checks conditional on storage backend
  so PostgreSQL mode doesn't report false errors in k8s (#2, #12)
- Cache listBackups() call to avoid duplicate disk I/O (#5)
- Rename loadConfigFromDisk to loadConfig since it reads from
  the repository abstraction, not necessarily disk (#3)
- Add assertInitialized() guard to db-store to catch usage
  before initDb() is called (#4)
- Extract dashboard CSS/JS to static files served via
  /static/dashboard.css and /static/dashboard.js routes (#6)
- Export stopPendingSessionReaper and call it on shutdown (#7)
- Remove unused LONG_TIMER_MS constant from types.ts (#9)
- Unify duplicated PersistedResponsesStore type (#10)
- Fix XSS in onboarding error display using textContent (#11)
- Update tests to call initDb() in setup where needed
