+++ src/types.ts
@@ -640,5 +640,3 @@ export const REQUEST_CLIENT_METADATA =
     platform: "MACOS",
     pluginType: "GEMINI",
   });
-
-export const LONG_TIMER_MS = 7 * 24 * 60 * 60 * 1000; // 7 days
